requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
numpy
opencv-python-headless
Pillow
pandas
requests
PyYAML
tqdm
matplotlib
seaborn
scipy
ultralytics
psutil
thop
gitpython
packaging
setuptools
fastapi
uvicorn
python-multipart
